Split wrapper.spec into separate files per top-level describe
